reorganizing modules to prepare for test config schema addition - simplifying structure and making more intuitive to find files
